+++ src/kernel/filesystem/vfs_layer/vfs_directory.h
@@ -15,7 +15,7 @@ public:
 	VFS_Directory(std::shared_ptr<Fat_Directory_Factory> factory, char file_name[12], std::uint8_t file_attributes);
 
 	//VFS ELEMENT METHODS
-	virtual void Open(std::uint16_t file_start, std::uint16_t file_size);
+	virtual void Open(std::uint16_t file_start, std::uint64_t file_size);
 	virtual bool Create();
 	virtual bool Remove();
 	virtual std::uint64_t Write(size_t how_many_bytes, void* buffer);
