+++ src/kernel/filesystem/vfs_layer/vfs_file.h
@@ -15,7 +15,7 @@ protected:
 public:
 	VFS_File(std::shared_ptr<Fat_File_Factory> fat_file_factory, std::shared_ptr<Fat_Directory> parent_directory, char file_name[12], std::uint8_t file_attributes);
 
-	virtual void Open(std::uint16_t file_start, std::uint16_t file_size);
+	virtual void Open(std::uint16_t file_start, std::uint64_t file_size);
 	virtual bool Create();
 	virtual bool Remove();
